requirements.txt
click==8.0.3
Flask==2.0.2
Flask-CKEditor==0.4.6
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
psycopg2-binary==2.9.1
python-dotenv==0.19.1
python-slugify==5.0.2
SQLAlchemy==1.4.26
text-unidecode==1.3
Werkzeug==2.0.2
WTForms==3.0.0
